JIT: move indirect call transformations to a new source file
Follow-up from a review comment in #21270.
